% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_helper_functions.R
\name{apply_sparseMatrix}
\alias{apply_sparseMatrix}
\title{slam private function for matrix apply function}
\usage{
apply_sparseMatrix(X, MARGIN, FUN, ...)
}
\description{
slam private function for matrix apply function
}
\keyword{internal}
