% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_de_sim_func.R
\name{sc_de_sim_func}
\alias{sc_de_sim_func}
\title{sc_de_sim_func.R}
\usage{
sc_de_sim_func(
  in_dir,
  gene_info_file,
  gene_name_column,
  meta_file,
  sample_column,
  cell_type_column,
  cell_barcode_column,
  contrast_column,
  contrast_ref_group,
  contrast_alt_group,
  count_file,
  cluster,
  covars,
  simulation_type = "de",
  fc_estimate = FALSE,
  simulation_mode = 3,
  fc,
  pDE,
  nSim.subjects,
  n_simulation = 1,
  main_seed,
  percentage.cell = 1,
  minimum.cells.per.gene.type = "or",
  minimum.cells.per.subject = 5,
  sim_cores = 5,
  de_methods = c("ancova", "t_test", "u_test", "edgeR", "limma", "DESeq2.shrink",
    "DESeq2", "DEseq2.nofilt", "nebula_LN", "nebula_HL", "glmmTMB"),
  sim_seed_mode = "nsim_mode",
  simulation_seeds
)
}
\arguments{
\item{in_dir}{The directory where the input files are located in 10X cellranger v2 format}

\item{gene_info_file}{The filename of the gene info file, csv format}

\item{meta_file}{The filename of the cell metadata file, csv format}

\item{sample_column}{The name of the sample column in the metadata file}

\item{cell_type_column}{The name of the cell type in the metadata file}

\item{cell_barcode_column}{The name of the cell barcode indices column in the metadata file}

\item{contrast_column}{The name of the DE contrast column in the metadata file, needs 2 levels, needs to map to samples}

\item{contrast_ref_group}{The level of the contrast to use as a reference}

\item{contrast_alt_group}{The level of the contrast to use as an alternative group}

\item{count_file}{The name of the count file (matrix market format)}

\item{cluster}{The name of the cell type to use for the contrast}

\item{covars}{The covariates in the metadata file}

\item{simulation_type}{Simulation type to run.  Values allowed are 'de' and 'null'.  Default is 'de'.}

\item{fc_estimate}{If running a when \code{simulation_type} is set to 'de', set \code{fc_estimate} to TRUE to run.  Default is FALSE.}

\item{simulation_mode}{Change the way the simulation is performed.  Values allowed are 1,2, and 3.  Default is 3.}

\item{fc}{The simulated fold change.  When \code{simulation_type} set to 'null' then simulations \code{fc} is set to 1.}

\item{pDE}{The percentage of genes that will be DE in the simulation}

\item{nSim.subjects}{The number of subjects to simulate in each contrast group}

\item{n_simulation}{The number of simulations to perform.  Default is 1.}

\item{main_seed}{Set a seed for reproducible simulations.  This seed is used in both modes of \code{sim_seed_mode}.}

\item{percentage.cell}{Percentage of cells to downsample for use in the simulation.  Useful for cell types with large numbers of cells.  Default is 1, i.e. no downsampling.}

\item{minimum.cells.per.gene.type}{Contrast-level filtering.  Values allowed are 'or' and 'and'.  Default is 'or'.  See public method apply_filter_contrasts_R6 in BiostatsSingleCell for details.}

\item{minimum.cells.per.subject}{Contrast-level filtering.  Specifies minimum number of cells per subject to retain.  Default is 5.  See public method apply_filter_contrasts_R6 in BiostatsSingleCell for details.}

\item{sim_cores}{Number of cores to use to run the simulations.  Default is 5.}

\item{de_methods}{DE methods to run from. Allowed values include 'ancova', 't_test', 'u_test', 'edgeR', 'limma', 'DESeq2.shrink', 'DESeq2', 'DEseq2.nofilt', 'nebula_LN', 'nebula_HL', 'glmmTMB', 'MAST'
Please note when including 'MAST' in de_methods, you will only be able to run one simulation at a time in either mode.
Default value is all methods excluding MAST.}

\item{sim_seed_mode}{Mode to use for seed.  'nsim_mode' will generate seeds based on the \code{n_simulation}.  'seed_mode' will run the seeds specified by the user by \code{simulation_seeds}.
Default value is 'nsim_mode'.}

\item{simulation_seeds}{A vector of simulation seeds to run.}
}
\value{
A dataframe of relevant simulation data (FC bias, null alpha, or DE performance parameters)
}
\description{
Generate Single Cell DE Simulation Results
}
\details{
This function runs a single cell DE simulation to obtain DE method performance for DE methods

The \code{simulation_mode} argument runs with three different modes
\itemize{
\item Mode 1 : Simulate mean/disp. in ref and alt groups based on mean/disp. from real data ref and alt groups, respectively.
\item Mode 2 : Simulate ref mean/disp. from real ref data; derive alt mean estimate using real ref mean multiplied by simulation FC;
derive alt disp. estimate either from real alt data for abundant genes or from median disp. for lowly-expressed genes.
\item Mode 3 : Simulate mean/disp. from real ref data; derive alt mean estimate using real ref mean multiplied by simulation FC;
predict alt disp. by testing simulation alt. mean in general additive model (GAM) after training the GAM on disp. ~ mean from real data
}
}
