% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_helper_functions.R
\name{sweep_sparseMatrix}
\alias{sweep_sparseMatrix}
\title{slam private function for matrix sweep function}
\usage{
sweep_sparseMatrix(x, margin, stats, fun = "*")
}
\description{
slam private function for matrix sweep function
}
\keyword{internal}
